Simplified remove_group()
Since we already fill the stack with stones of the targeted group in
group_has_liberties() there's no need to search for the group again.
This requires modifying group_has_liberties() so it can return the group
if it is needed, which does make the interface of the function a bit
confusing but the only alternative would be removing the stones in the
same function and naming such a function would be a nightmare of its
own, since it is also used to just check it there are liberties left.